@@ -63,15 +63,14 @@
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-		<java.version>11</java.version>
+		<java.version>17</java.version>
 		<maven.compiler.source>${java.version}</maven.compiler.source>
 		<maven.compiler.target>${java.version}</maven.compiler.target>
 		<xtext.version>2.30.0</xtext.version>
 		<junit.version>5.7.2</junit.version>
 		<lombok.version>1.18.26</lombok.version>
 		<spotify.hamcrest.version>1.3.1</spotify.hamcrest.version>
-		<springboot.version>2.7.10</springboot.version>
-		<springsecurity.version>6.0.2</springsecurity.version>
+		<springboot.version>3.0.5</springboot.version>
 		<vaadin.version>24.0.2</vaadin.version>
 		<reactornetty.version>1.0.23</reactornetty.version>
 		<rsocket.version>1.1.3</rsocket.version>
@@ -100,6 +99,7 @@
 		<module>sapl-rsocket-endpoint</module>
 		<module>sapl-documentation</module>
 		<module>sapl-server-lt</module>
+		<module>sapl-editor-vaadin</module>
 		<module>sapl-bom</module>
 	</modules>
 	<dependencyManagement>
@@ -138,6 +138,14 @@
 				<groupId>org.springframework.boot</groupId>
 				<artifactId>spring-boot-dependencies</artifactId>
 				<version>${springboot.version}</version>
+				<type>pom</type>
+				<scope>import</scope>
+			</dependency>
+			<!-- Vaadin -->
+			<dependency>
+				<groupId>com.vaadin</groupId>
+				<artifactId>vaadin-bom</artifactId>
+				<version>${vaadin.version}</version>
 				<type>pom</type>
 				<scope>import</scope>
 			</dependency>
