@@ -15,7 +15,7 @@
  */
 package io.sapl.mavenplugin.test.coverage.report.sonar.model;
 
-import javax.xml.bind.annotation.XmlRegistry;
+import jakarta.xml.bind.annotation.XmlRegistry;
 
 /**
  * This object contains factory methods for each Java content interface and Java element
