@@ -24,7 +24,6 @@ import java.util.Queue;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentLinkedQueue;
 
-import org.springframework.http.HttpStatus;
 import org.springframework.web.reactive.function.client.ClientResponse;
 import org.springframework.web.reactive.function.client.WebClient;
 import org.springframework.web.reactive.function.client.WebClient.ResponseSpec;
@@ -136,7 +135,7 @@ public class JWTKeyProvider {
 			response = webClient.get().uri(publicKeyURI, kid).retrieve();
 		}
 
-		return response.onStatus(HttpStatus::isError, this::handleHttpError).bodyToMono(String.class)
+		return response.onStatus(status -> status.isError(), this::handleHttpError).bodyToMono(String.class)
 				.map(JWTEncodingDecodingUtils::encodedX509ToRSAPublicKey).filter(Optional::isPresent)
 				.map(Optional::get);
 	}
