@@ -20,14 +20,14 @@ import java.util.Enumeration;
 import java.util.Locale;
 import java.util.Map.Entry;
 
-import javax.servlet.http.Cookie;
-import javax.servlet.http.HttpServletRequest;
-
 import org.springframework.boot.jackson.JsonComponent;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.JsonSerializer;
 import com.fasterxml.jackson.databind.SerializerProvider;
+
+import jakarta.servlet.http.Cookie;
+import jakarta.servlet.http.HttpServletRequest;
 
 @JsonComponent
 public class HttpServletRequestSerializer extends JsonSerializer<HttpServletRequest> {
